requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
PyJWT==2.6.0
bcrypt==3.2.0
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.5